// web/packages/elena/src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      controllers: {
        Row: {
          current_profile: string | null
          current_temp: number
          id: string
          is_running: boolean
          last_updated: string
          max_temp: number
          min_temp: number
          name: string
          slave_id: number
          target_temp: number
          update_interval: number
          zone_id: string
        }
        Insert: {
          current_profile?: string | null
          current_temp: number
          id?: string
          is_running?: boolean
          last_updated?: string
          max_temp: number
          min_temp: number
          name: string
          slave_id: number
          target_temp: number
          update_interval: number
          zone_id: string
        }
        Update: {
          current_profile?: string | null
          current_temp?: number
          id?: string
          is_running?: boolean
          last_updated?: string
          max_temp?: number
          min_temp?: number
          name?: string
          slave_id?: number
          target_temp?: number
          update_interval?: number
          zone_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "controllers_current_profile_fkey"
            columns: ["current_profile"]
            isOneToOne: false
            referencedRelation: "temperature_profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "controllers_zone_id_fkey"
            columns: ["zone_id"]
            isOneToOne: false
            referencedRelation: "heat_zones"
            referencedColumns: ["id"]
          },
        ]
      }
      heat_zones: {
        Row: {
          description: string | null
          id: string
          name: string
        }
        Insert: {
          description?: string | null
          id?: string
          name: string
        }
        Update: {
          description?: string | null
          id?: string
          name?: string
        }
        Relationships: []
      }
      temperature_profiles: {
        Row: {
          control_points: Json
          created_at: string
          description: string | null
          duration: number
          id: string
          name: string
          updated_at: string
        }
        Insert: {
          control_points: Json
          created_at?: string
          description?: string | null
          duration: number
          id?: string
          name: string
          updated_at?: string
        }
        Update: {
          control_points?: Json
          created_at?: string
          description?: string | null
          duration?: number
          id?: string
          name?: string
          updated_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never